recorder files, commands and functions
