--- daily_report.package.json
@@ -1,18 +1,0 @@
-{
-  "name": "aws-cost-explorer",
-  "version": "1.0.0",
-  "description": "",
-  "main": "daily_report.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "async": "3.1.0",
-    "aws-sdk": "2.518.0",
-    "csv-parse": "4.4.5",
-    "request": "2.88.0",
-    "unzipper": "0.10.0"
-  }
-}
