+++ package.json
@@ -4,14 +4,17 @@
   "description": "",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "standard": "node_modules/.bin/standard \"**.js\" --parser babel-eslint"
+    "standard": "node_modules/.bin/standard \"**.js\" --parser babel-eslint",
+    "standard:fix": "node_modules/.bin/standard \"**.js\" --parser babel-eslint --fix",
+    "report:daily": "node src/daily_report.js",
+    "report:monthly": "node src/daily_report.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "async": "3.1.0",
     "aws-sdk": "2.518.0",
-    "csv-parse": "4.4.5",
+    "csv-parse": "^4.11.1",
     "request": "2.87.0",
     "unzipper": "0.10.0"
   },
